+++ petbook_fe/components/register/RegisterForm.tsx
@@ -1,5 +1,5 @@
 import { useRecoilValue } from "recoil";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import Image from "next/image";
 
 // custom
@@ -11,7 +11,7 @@ import navigator from "@lib/modules/navigator";
 import RegisterInputBox from "@components/register/RegisterInputBox";
 import PasswordInput from "@components/register/RegisterPasswordFrom";
 import RegisterNameForm from "@components/register/RegisterNameForm";
-import { Modal } from "@components/common/Modal/Modal";
+import RegisterTerms from "@components/register/RegisterTerms";
 
 // type
 import { ErrorResponse } from "@lib/API/petBookAPI/types/userRequest";
@@ -19,7 +19,6 @@ import { ErrorResponse } from "@lib/API/petBookAPI/types/userRequest";
 // styled
 import {
   Main,
-  Terms,
   RegisterFormWrap,
   SpaceTopWrap,
   RegisterInfoText,
@@ -87,77 +86,6 @@ const Register = () => {
     </div>
   );
 };
-const TermsWrap = () => {
-  const [modal, setModal] = useState({
-    state: false,
-    data: {
-      content: "",
-      title: "",
-    },
-  });
-
-  const handleCloseModal = () => {
-    setModal((modalState) => ({ ...modalState, state: false }));
-  };
-
-  const HandlerOpen = (title: string) => {
-    let titleName = "";
-    let contentText = "";
-    switch (title) {
-      case "terms": {
-        titleName = "이용약관";
-        contentText = "약관 내용 // 지금은 임시";
-        break;
-      }
-      case "privacy": {
-        titleName = "개인정보 수집 및 이용약관";
-        contentText = "개인정보 수집 및 이용약관 내용 // 지금은 임시";
-        break;
-      }
-      default: {
-        break;
-      }
-    }
-    setModal((modalState) => ({
-      ...modalState,
-      state: true,
-      data: {
-        title: titleName,
-        content: contentText,
-      },
-    }));
-  };
-
-  return (
-    <Terms>
-      <Modal modalState={modal} handleCloseModal={handleCloseModal}>
-        <hgroup>
-          <h2>{modal.data.title}</h2>
-        </hgroup>
-        <section>
-          <p>{modal.data.content}</p>
-        </section>
-      </Modal>
-      <li>
-        <button type="button" onClick={() => HandlerOpen("terms")}>
-          <label htmlFor="terms">
-            <input type="checkbox" id="terms" />
-            <p>펫북 이용 약관에 동의합니다 [필수]</p>
-          </label>
-        </button>
-      </li>
-      <li>
-        <button type="button" onClick={() => HandlerOpen("privacy")}>
-          <label htmlFor="privacy">
-            <input type="checkbox" id="privacy" />
-            <p>개인정보 수집 및 이용에 동의합니다 [필수]</p>
-          </label>
-        </button>
-      </li>
-    </Terms>
-  );
-};
-
 const RegisterButton = () => {
   const registerForm = useRecoilValue(registerFormState);
   const { data, isSuccess, isError, error, mutate } =
@@ -184,7 +112,7 @@ const RegisterButton = () => {
 };
 
 RegisterContainer.Register = Register;
-RegisterContainer.TermsWrap = TermsWrap;
+RegisterContainer.TermsWrap = RegisterTerms;
 RegisterContainer.RegisterButton = RegisterButton;
 
 export default RegisterContainer;
